Move Plane code into its own files. Make PlaneClassification an enum for performance
