+++ simulant/generic/optional.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <cassert>
+#include <utility>
 
 namespace smlt {
