+++ tests/test_plane.h
@@ -23,15 +23,15 @@ public:
         smlt::Vec3 behind(0, -1, 0);
         smlt::Vec3 on_plane(0, 0, 0);
 
-        assert_true(p.classify_point(in_front).is_in_front_of_plane());
-        assert_false(p.classify_point(in_front).is_behind_plane());
+        assert_true(p.classify_point(in_front) == smlt::PLANE_CLASSIFICATION_IS_IN_FRONT_OF_PLANE);
+        assert_false(p.classify_point(in_front) == smlt::PLANE_CLASSIFICATION_IS_BEHIND_PLANE);
 
-        assert_true(p.classify_point(behind).is_behind_plane());
-        assert_false(p.classify_point(behind).is_in_front_of_plane());
+        assert_true(p.classify_point(behind) == smlt::PLANE_CLASSIFICATION_IS_BEHIND_PLANE);
+        assert_false(p.classify_point(behind) == smlt::PLANE_CLASSIFICATION_IS_IN_FRONT_OF_PLANE);
 
-        assert_true(p.classify_point(on_plane).is_on_plane());
-        assert_false(p.classify_point(on_plane).is_behind_plane());
-        assert_false(p.classify_point(on_plane).is_in_front_of_plane());
+        assert_true(p.classify_point(on_plane) == smlt::PLANE_CLASSIFICATION_IS_ON_PLANE);
+        assert_false(p.classify_point(on_plane) == smlt::PLANE_CLASSIFICATION_IS_BEHIND_PLANE);
+        assert_false(p.classify_point(on_plane) == smlt::PLANE_CLASSIFICATION_IS_IN_FRONT_OF_PLANE);
     }
 };
